feat swagger

+++ internal/handlers/favorites.go
@@ -22,6 +22,18 @@ type addMovieReq struct {
 	MovieID int `json:"movie_id"`
 }
 
+// AddFavoriteMovieToUser godoc
+// @Summary Добавить фильм в избранные
+// @Description Добавляет выбранный фильм в список избранных текущего пользователя.
+// @Tags Favorites
+// @Accept json
+// @Security BearerAuth
+// @Param request body addMovieReq true "Данные для добавления фильма"
+// @Success 201
+// @Failure 400 {object} map[string]string
+// @Failure 401 {object} map[string]string
+// @Failure 500 {object} map[string]string
+// @Router /favorites [post]
 func (h *FavoriteHandler) AddFavoriteMovieToUser(w http.ResponseWriter, r *http.Request) {
 	userID, ok := r.Context().Value(middleware.UserIDKey).(int)
 	if !ok {
@@ -36,6 +48,11 @@ func (h *FavoriteHandler) AddFavoriteMovieToUser(w http.ResponseWriter, r *http.
 		return
 	}
 
+	if req.MovieID <= 0 {
+		http.Error(w, "ID фильма должен быть выше 0", http.StatusBadRequest)
+		return
+	}
+
 	err := h.service.AddFavoriteMovieToUser(r.Context(), userID, req.MovieID)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
@@ -45,6 +62,17 @@ func (h *FavoriteHandler) AddFavoriteMovieToUser(w http.ResponseWriter, r *http.
 	w.WriteHeader(http.StatusCreated)
 }
 
+// GetFavoriteMovies godoc
+// @Summary Получить список избранных фильмов
+// @Description Возвращает список всех избранных фильмов текущего пользователя
+// @Tags Favorites
+// @Produce json
+// @Security BearerAuth
+// @Success 200 {array} models.FavoriteMovieResponse
+// @Failure 401 {object} map[string]string
+// @Failure 404 {object} map[string]string
+// @Failure 500 {object} map[string]string
+// @Router /favorites [get]
 func (h *FavoriteHandler) GetFavoriteMovies(w http.ResponseWriter, r *http.Request) {
 	userID, ok := r.Context().Value(middleware.UserIDKey).(int)
 	if !ok {
@@ -71,6 +99,18 @@ func (h *FavoriteHandler) GetFavoriteMovies(w http.ResponseWriter, r *http.Reque
 	}
 }
 
+// DeleteFavoriteMovieFromUser godoc
+// @Summary Удалить фильм из избранного
+// @Description Удаляет выбранный фильм из списка избранных текущего пользователя.
+// @Tags Favorites
+// @Security BearerAuth
+// @Param movie_id path int true "ID фильма"
+// @Success 204
+// @Failure 400 {object} map[string]string
+// @Failure 401 {object} map[string]string
+// @Failure 404 {object} map[string]string
+// @Failure 500 {object} map[string]string
+// @Router /favorites/{movie_id} [delete]
 func (h *FavoriteHandler) DeleteFavoriteMovieFromUser(w http.ResponseWriter, r *http.Request) {
 	userID, ok := r.Context().Value(middleware.UserIDKey).(int)
 	if !ok {
@@ -78,16 +118,21 @@ func (h *FavoriteHandler) DeleteFavoriteMovieFromUser(w http.ResponseWriter, r *
 		return
 	}
 
-	movieIDstr := r.PathValue("movie_id")
-	movieID, err := strconv.Atoi(movieIDstr)
+	movieIDStr := r.PathValue("movie_id")
+	movieID, err := strconv.Atoi(movieIDStr)
 	if err != nil {
 		http.Error(w, "Неверный ID фильма", http.StatusBadRequest)
 		return
 	}
 
+	if movieID <= 0 {
+		http.Error(w, "ID должен быть выше 0", http.StatusBadRequest)
+		return
+	}
+
 	err = h.service.DeleteFavoriteMovieFromUser(r.Context(), userID, movieID)
 	if err != nil {
-		if errors.Is(err, repositories.ErrMovieNotFound) {
+		if errors.Is(err, repositories.ErrFavoriteNotFound) {
 			http.Error(w, err.Error(), http.StatusNotFound)
 			return
 		}
